BMesh intersect: Add function to check whether the point is inside the mesh
It gets the BVH tree and loop triangles as an input and checks whether the
given point is inside of the mesh represented by those data or not.

It is not optimized for speed yet and could have some numerical instabilities
which we'll improve later.

P.S. This function is currently unused.